WEB-3919: separate out pro upgrade prompt component to own file
