+++ app/(candidate)/dashboard/shared/ProUpgradeModal.js
@@ -1,120 +1,16 @@
 'use client'
 
-import { useState, useEffect } from 'react'
 import Modal from '@shared/utils/Modal'
 import H1 from '@shared/typography/H1'
 import Body2 from '@shared/typography/Body2'
 import Button from '@shared/buttons/Button'
-import { trackEvent, EVENTS } from 'helpers/fullStoryHelper'
 
 export const VIABILITY_SCORE_THRESHOLD = 2
-const LOCAL_STORAGE_KEY = 'proUpgradeModalDismissedSession'
-
 export const VARIANTS = {
   First: 'First',
   Second_NonViable: 'Second_NonViable',
   Second_Viable: 'Second_Viable',
   Third: 'Third',
-}
-
-const SESSION_TRIGGERS = {
-  First: 3,
-  Second: 8,
-  Third: 12,
-}
-
-export function ProUpgradePrompt({ campaign, user }) {
-  const isPro = campaign?.isPro || false
-  const sessionCount = user?.metaData?.sessionCount || 0
-  const viablityScore = campaign?.pathToVictory?.data?.viability?.score || 0
-
-  const [modalState, setModalState] = useState({
-    isOpen: false,
-    variant: VARIANTS.First,
-  })
-
-  useEffect(() => {
-    if (isPro) {
-      setModalState((current) => ({ ...current, isOpen: false }))
-      return
-    }
-
-    const parsed = parseInt(localStorage.getItem(LOCAL_STORAGE_KEY))
-    const lastDismissedSession = Number.isNaN(parsed) ? 0 : parsed
-
-    if (
-      sessionCount >= SESSION_TRIGGERS.First &&
-      sessionCount < SESSION_TRIGGERS.Second &&
-      lastDismissedSession < SESSION_TRIGGERS.First
-    ) {
-      // first trigger
-      handleOpen(VARIANTS.First)
-    } else if (
-      sessionCount >= SESSION_TRIGGERS.Second &&
-      sessionCount < SESSION_TRIGGERS.Third &&
-      lastDismissedSession < SESSION_TRIGGERS.Second
-    ) {
-      // second trigger, show differnet content based on viability score
-      handleOpen(
-        viablityScore < VIABILITY_SCORE_THRESHOLD
-          ? VARIANTS.Second_NonViable
-          : VARIANTS.Second_Viable,
-      )
-    } else if (
-      sessionCount >= SESSION_TRIGGERS.Third &&
-      lastDismissedSession < SESSION_TRIGGERS.Third
-    ) {
-      // third trigger
-      handleOpen(VARIANTS.Third)
-    }
-
-    function handleOpen(variant) {
-      trackEvent(EVENTS.ProUpgrade.Modal.Shown, {
-        sessionCount,
-        viablityScore,
-        variant,
-      })
-      setModalState({
-        isOpen: true,
-        variant: variant,
-      })
-    }
-  }, [sessionCount, viablityScore, isPro])
-
-  // Don't want to show modal if campaign is already pro
-  if (isPro) return null
-
-  function handleClose() {
-    trackEvent(EVENTS.ProUpgrade.Modal.Exit, {
-      sessionCount,
-      viablityScore,
-      variant: modalState.variant,
-    })
-    closeModal()
-  }
-
-  function handleUpgradeLinkClick() {
-    trackEvent(EVENTS.ProUpgrade.Modal.ClickButton, {
-      sessionCount,
-      viablityScore,
-      variant: modalState.variant,
-    })
-    closeModal()
-  }
-
-  function closeModal() {
-    localStorage.setItem(LOCAL_STORAGE_KEY, sessionCount)
-    setModalState((current) => ({ ...current, isOpen: false }))
-  }
-
-  return (
-    <ProUpgradeModal
-      open={modalState.isOpen}
-      variant={modalState.variant}
-      onClose={handleClose}
-      onUpgradeLinkClick={handleUpgradeLinkClick}
-    />
-  )
 }
 
 export function ProUpgradeModal({
